Removed Vuex Restructured Store folder Uninstalled Vuex package
