+++ package.json
@@ -21,8 +21,7 @@
     "pinia": "^2.0.23",
     "tachyons": "^4.12.0",
     "vue": "^3.2.25",
-    "vue-router": "^4.0.12",
-    "vuex": "^4.0.2"
+    "vue-router": "^4.0.12"
   },
   "devDependencies": {
     "@babel/core": "^7.18.10",
