+++ src/components/Nav/Nav.spec.ts
@@ -1,7 +1,7 @@
 import { shallowMount, VueWrapper } from '@vue/test-utils';
 
 import { createTestingPinia } from '@pinia/testing';
-import { useUserStore } from '@/store/modules/user';
+import { useUserStore } from '@/store/user';
 import { useNavigation } from '@/modules/navigation';
 
 import Nav from './Nav.vue';
