+++ src/components/Nav/Nav.vue
@@ -1,6 +1,6 @@
 <script setup lang="ts">
 import { computed } from 'vue';
-import { useUserStore } from '@/store/modules/user'
+import { useUserStore } from '@/store/user';
 import { useNavigation } from '@/modules/navigation';
 import { Routes } from '@/router/routes';
 import ProfileIcon from '../Profile/ProfileIcon.vue';
