+++ src/main.ts
@@ -1,7 +1,6 @@
 import { createApp } from 'vue';
 import { createPinia } from 'pinia';
 import App from './App.vue';
-import store from './store';
 import router from './router';
 
 import Particles from 'particles.vue3';
@@ -11,4 +10,4 @@ import 'bootstrap';
 
 const pinia = createPinia();
 
-createApp(App).use(Particles).use(pinia).use(store).use(router).mount('#app');
+createApp(App).use(Particles).use(pinia).use(router).mount('#app');
