+++ src/store/image/image.spec.ts
