+++ src/store/image/imageTypes.ts
