+++ src/store/image/index.ts
