--- src/store/rootTypes.ts
@@ -1,3 +1,0 @@
-/* eslint-disable @typescript-eslint/ban-types */
-export type RootState = {};
-export type RootGetters = Record<string, any>;
