+++ src/views/Register/Register.spec.ts
@@ -1,7 +1,7 @@
 import { describe, expect, it, beforeEach, vi, beforeAll } from 'vitest';
 import { shallowMount, VueWrapper } from '@vue/test-utils';
 import { createTestingPinia } from '@pinia/testing';
-import { useUserStore } from '@/store/modules/user';
+import { useUserStore } from '@/store/user';
 import { useNavigation } from '@/modules/navigation';
 import { Routes } from '@/router/routes';
